requirements.txt
Pillow==6.2.1	
PyAudio==0.2.11	
SpeechRecognition==3.8.1	
gTTS==2.1.1	
gTTS-token==1.1.3	
googletrans==3.0.0	
pytesseract==0.3.0	
pyttsx3==2.71	
tesseract==0.1.3	
wheel==0.34.2	
